# docs/swagger.yaml
basePath: /
definitions:
  api.CityResponse:
    properties:
      county:
        type: string
      province:
        type: string
    type: object
...
  api.cancelReservationResponse:
    properties:
      amount_refunded:
        type: integer
      change_reservation:
        $ref: '#/definitions/db.ChangeReservation'
      message:
        type: string
      ticket_id:
        type: integer
    type: object
...
  api.searchTicketsRequest:
    properties:
      departure_date:
        type: string
      destination_city_id:
        type: integer
      origin_city_id:
        type: integer
      vehicle_type:
        enum:
        - BUS
        - TRAIN
        - AIRPLANE
        type: string
    required:
    - departure_date
    - destination_city_id
    - origin_city_id
    - vehicle_type
    type: object
...
  db.ChangeReservation:
    properties:
      admin_id:
        type: integer
      from_status:
        $ref: '#/definitions/db.TicketStatus'
      id:
        type: integer
      reservation_id:
        type: integer
      to_status:
        $ref: '#/definitions/db.TicketStatus'
      user_id:
        type: integer
    type: object
  db.CheckReservationTicketStatus:
    enum:
    - RESERVED
    - NOT_RESERVED
    type: string
    x-enum-varnames:
    - CheckReservationTicketStatusRESERVED
    - CheckReservationTicketStatusNOTRESERVED
...
  db.Penalty:
    properties:
      after_day:
        type: integer
      befor_day:
        type: integer
      id:
        type: integer
      penalty_text:
        type: string
      vehicle_id:
        type: integer
    type: object
...
  db.Ticket:
    properties:
      amount:
        type: integer
      arrival_time:
        type: string
      count_stand:
        type: integer
      created_at:
        type: string
      departure_time:
        type: string
      id:
        type: integer
      route_id:
        type: integer
      seat_id:
        type: integer
      status:
        $ref: '#/definitions/db.CheckReservationTicketStatus'
      vehicle_id:
        type: integer
      vehicle_type:
        $ref: '#/definitions/db.VehicleType'
    type: object
  db.TicketStatus:
    enum:
    - RESERVED
    - RESERVING
    - CANCELED
    - CANCELED-BY-TIME
    type: string
    x-enum-varnames:
    - TicketStatusRESERVED
    - TicketStatusRESERVING
    - TicketStatusCANCELED
    - TicketStatusCANCELEDBYTIME
...
  db.VehicleType:
    enum:
    - BUS
    - TRAIN
    - AIRPLANE
    type: string
    x-enum-varnames:
    - VehicleTypeBUS
    - VehicleTypeTRAIN
    - VehicleTypeAIRPLANE
host: localhost:3000
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: API for Bilitio ticket booking system.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://localhost:3000/swagger/index.html
  title: Bilitio API
  version: "1.0"
paths:
...
  /city:
    get:
      consumes:
      - application/json
      description: Returns a list of all cities available for booking.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.CityResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all cities
      tags:
      - Cities
...
  /penalty/{ticket_id}:
    put:
      consumes:
      - application/json
      description: Cancel a ticket reservation and calculate penalty refund. Requires
        authentication.
      parameters:
      - description: Ticket ID
        in: path
        name: ticket_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.cancelReservationResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cancel a reservation
      tags:
      - Reservation
...
  /ticket-penalties/{ticket_id}:
    get:
      consumes:
      - application/json
      description: Retrieve penalties associated with a specific ticket. Requires
        authentication.
      parameters:
      - description: Ticket ID
        in: path
        name: ticket_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of penalties
          schema:
            items:
              $ref: '#/definitions/db.Penalty'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get penalties for a ticket
      tags:
      - Penalties
...
  /tickets/search:
    post:
      consumes:
      - application/json
      description: Search tickets by origin, destination, date, and vehicle type
      parameters:
      - description: Search tickets request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.searchTicketsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.Ticket'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Search tickets
      tags:
      - tickets
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"